Knexseeds

+++ migrations/20180616122920_users.js
@@ -5,7 +5,7 @@ exports.up = function(knex, Promise) {
 
     table.varchar('first_name', 50).notNullable().defaultTo('')
     table.varchar('last_name', 50).notNullable().defaultTo('')
-    table.integer('phone_number').notNullable().defaultTo(null)
+    table.bigInteger('phone_number').notNullable().defaultTo(null)
     table.varchar('email', 50).notNullable().defaultTo('')
     table.varchar('password', 50).notNullable().defaultTo('')
     table.boolean('ispm').notNullable().defaultTo(false)
